Little changes that fixes a compiler warnings.
- changes in an enum type definition (enum -> NS_ENUM)
- changes the property named description in the EEUsersQuery and EEUser object  to desc.

- fixes NSInteger conversion to long
- fixes conversion between the NSMutableDictionary and NSDictionary types

Enable The following recommended compiler warnings:
- Empty Loop Bodies
- Implicit Boolean Conversions
- Implicit Constant Conversions
- Implicit Conversion to 32 Bit Type
- Implicit Enum Conversions
- Implicit Integer to Pointer Conversions
- Undeclared Selector
- Unused Functions
- Enable Strict Checking of objc_msgSend Calls
- Unreachable Code

and:
- Automatically Select Architectures
- Enable Build Active Architecture Only When Debugging

+++ EEAPI/EyeEmAPI.m
@@ -395,7 +395,7 @@
         return nil;
     }
     NSString *newUrl = [self extractString:[urlPath absoluteString]];
-    newUrl = [NSString stringWithFormat:@"%@/%i/%i/%@", kURL, (NSInteger) roundf(imageSize.width), (NSInteger) roundf(imageSize.height), newUrl];
+    newUrl = [NSString stringWithFormat:@"%@/%li/%li/%@", kURL, (long)roundf(imageSize.width), (long)roundf(imageSize.height), newUrl];
     
     
     return [NSURL URLWithString:newUrl];
@@ -408,7 +408,7 @@
         return nil;
     }
     NSString *newUrl = [self extractString: [urlPath absoluteString]];
-    newUrl = [NSString stringWithFormat:@"%@/w/%i/%@", kURL, (imageWidth), newUrl];
+    newUrl = [NSString stringWithFormat:@"%@/w/%li/%@", kURL, (long)(imageWidth), newUrl];
     
     
     return [NSURL URLWithString:newUrl];
@@ -419,7 +419,7 @@
         return nil;
     }
     NSString *newUrl = [self extractString:[urlPath absoluteString]];
-    newUrl = [NSString stringWithFormat:@"%@/sq/%i/%@", kURL, (imageSize), newUrl];
+    newUrl = [NSString stringWithFormat:@"%@/sq/%li/%@", kURL, (long)(imageSize), newUrl];
     
     
     return [NSURL URLWithString:newUrl];
@@ -432,7 +432,7 @@
         return nil;
     }
     NSString *newUrl = [self extractString:[urlPath absoluteString]];
-    newUrl = [NSString stringWithFormat:@"%@/h/%i/%@", kURL, (imageHeight), newUrl];
+    newUrl = [NSString stringWithFormat:@"%@/h/%li/%@", kURL, (long)(imageHeight), newUrl];
     
     
     return [NSURL URLWithString:newUrl];
